update filter file name, separate requirements constant into separate file
