+++ src/tests/unit/tests/assessments/assessments-requirements-filter.test.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { assessmentsProviderForRequirements } from 'assessments/assessments-medium-pass-filter';
+import { assessmentsProviderForRequirements } from 'assessments/assessments-requirements-filter';
 import { AssessmentsProviderImpl } from 'assessments/assessments-provider';
 import { AutomatedChecks } from 'assessments/automated-checks/assessment';
 import { AssessmentsProvider } from 'assessments/types/assessments-provider';
